refactor: extract methods to dedicated files per review (ObsoleteSeverity->Enums, CaptureConnections, RestoreConnections, CopyPersistentData, ExpandVariableParameters, CopyPersistentDataFromComponent, ParamConnection)
