+++ Grasshopper/SAM.Core.Grasshopper/Enums/ObsoleteSeverity.cs
