+++ Grasshopper/SAM.Core.Grasshopper/Modify/CopyParameters.cs
@@ -2,7 +2,6 @@
 // Copyright (c) 2020–2026 Michal Dengusiak & Jakub Ziolkowski and contributors
 
 using Grasshopper.Kernel;
-using System;
 using System.Collections.Generic;
 using System.Linq;
 
@@ -10,13 +9,6 @@ namespace SAM.Core.Grasshopper
 {
     public static partial class Modify
     {
-        public class ParamConnection
-        {
-            public GH_ParameterSide Side;
-            public string ParamName;
-            public List<IGH_Param> ConnectedParams = new List<IGH_Param>();
-        }
-
         public static void CopyParameters(GH_ParameterSide parameterSide, GH_SAMComponent gH_SAMComponent_From, GH_SAMComponent gH_SAMComponent_To)
         {
             if (gH_SAMComponent_From == null || gH_SAMComponent_To == null)
@@ -127,204 +119,6 @@ namespace SAM.Core.Grasshopper
                         }
                     }
                 }
-            }
-        }
-
-        public static List<ParamConnection> CaptureConnections(GH_SAMComponent gH_SAMComponent)
-        {
-            List<ParamConnection> result = new List<ParamConnection>();
-
-            if (gH_SAMComponent == null)
-            {
-                return result;
-            }
-
-            List<IGH_Param> outputParams = gH_SAMComponent.Params?.Output;
-            if (outputParams != null)
-            {
-                foreach (IGH_Param param in outputParams)
-                {
-                    if (param == null || param.Recipients == null || param.Recipients.Count == 0)
-                    {
-                        continue;
-                    }
-
-                    result.Add(new ParamConnection
-                    {
-                        Side = GH_ParameterSide.Output,
-                        ParamName = param.Name,
-                        ConnectedParams = new List<IGH_Param>(param.Recipients)
-                    });
-                }
-            }
-
-            List<IGH_Param> inputParams = gH_SAMComponent.Params?.Input;
-            if (inputParams != null)
-            {
-                foreach (IGH_Param param in inputParams)
-                {
-                    if (param == null || param.Sources == null || param.Sources.Count == 0)
-                    {
-                        continue;
-                    }
-
-                    result.Add(new ParamConnection
-                    {
-                        Side = GH_ParameterSide.Input,
-                        ParamName = param.Name,
-                        ConnectedParams = new List<IGH_Param>(param.Sources)
-                    });
-                }
-            }
-
-            return result;
-        }
-
-        public static void RestoreConnections(GH_SAMComponent gH_SAMComponent, List<ParamConnection> connections, out Log log)
-        {
-            log = new Log();
-
-            if (gH_SAMComponent == null || connections == null || connections.Count == 0)
-            {
-                return;
-            }
-
-            List<IGH_Param> outputParams = gH_SAMComponent.Params?.Output;
-            List<IGH_Param> inputParams = gH_SAMComponent.Params?.Input;
-
-            Dictionary<string, IGH_Param> newOutputParams = new Dictionary<string, IGH_Param>();
-            Dictionary<string, IGH_Param> newInputParams = new Dictionary<string, IGH_Param>();
-
-            if (outputParams != null)
-            {
-                foreach (IGH_Param param in outputParams)
-                {
-                    if (param != null)
-                    {
-                        newOutputParams[param.Name] = param;
-                    }
-                }
-            }
-
-            if (inputParams != null)
-            {
-                foreach (IGH_Param param in inputParams)
-                {
-                    if (param != null)
-                    {
-                        newInputParams[param.Name] = param;
-                    }
-                }
-            }
-
-            foreach (ParamConnection connection in connections)
-            {
-                Dictionary<string, IGH_Param> targetDict = connection.Side == GH_ParameterSide.Output ? newOutputParams : newInputParams;
-
-                if (!targetDict.TryGetValue(connection.ParamName, out IGH_Param newParam))
-                {
-                    if (connection.ConnectedParams.Count > 0)
-                    {
-                        string sideLabel = connection.Side == GH_ParameterSide.Output ? "output" : "input";
-                        log.Add(new LogRecord("  DROPPED: {0} {1} '{2}' not in new version ({3} {4} lost). Manual reconnect needed.",
-                            LogRecordType.Warning, gH_SAMComponent.Name, sideLabel, connection.ParamName,
-                            connection.ConnectedParams.Count, connection.ConnectedParams.Count == 1 ? "connection" : "connections"));
-
-                        foreach (IGH_Param cp in connection.ConnectedParams)
-                        {
-                            if (cp != null)
-                            {
-                                string dest = "?";
-                                try { dest = cp.Attributes?.ToString() ?? cp.Name; } catch { dest = cp.Name; }
-                                log.Add(new LogRecord("    was connected to: {0}", LogRecordType.Warning, dest));
-                            }
-                        }
-                    }
-                    continue;
-                }
-
-                foreach (IGH_Param connectedParam in connection.ConnectedParams)
-                {
-                    if (connectedParam == null)
-                    {
-                        continue;
-                    }
-
-                    try
-                    {
-                        if (connection.Side == GH_ParameterSide.Output)
-                        {
-                            connectedParam.AddSource(newParam);
-                        }
-                        else
-                        {
-                            newParam.AddSource(connectedParam);
-                        }
-                    }
-                    catch (Exception ex)
-                    {
-                        log.Add(new LogRecord("  Warning: failed to reconnect '{0}' → '{1}': {2}",
-                            LogRecordType.Warning, connection.ParamName,
-                            connectedParam.Name ?? "(unnamed)", ex.Message));
-                    }
-                }
-            }
-        }
-
-        internal static void CopyPersistentData(IGH_Param oldParam, IGH_Param newParam)
-        {
-            if (oldParam == null || newParam == null)
-            {
-                return;
-            }
-
-            if (oldParam.GetType() != newParam.GetType())
-            {
-                return;
-            }
-
-            try
-            {
-                Type type = oldParam.GetType();
-                while (type != null && type != typeof(object))
-                {
-                    if (type.IsGenericType && type.GetGenericTypeDefinition().FullName == "Grasshopper.Kernel.GH_PersistentParam`1")
-                    {
-                        System.Reflection.PropertyInfo persistentDataProp = type.GetProperty("PersistentData");
-                        if (persistentDataProp == null)
-                        {
-                            break;
-                        }
-
-                        object persistentData = persistentDataProp.GetValue(oldParam);
-                        if (persistentData == null)
-                        {
-                            break;
-                        }
-
-                        if (!(persistentData is System.Collections.IEnumerable enumerable))
-                        {
-                            break;
-                        }
-
-                        System.Reflection.MethodInfo addMethod = type.GetMethod("AddPersistentData");
-                        if (addMethod == null)
-                        {
-                            break;
-                        }
-
-                        foreach (object item in enumerable)
-                        {
-                            addMethod.Invoke(newParam, new[] { item });
-                        }
-                        break;
-                    }
-                    type = type.BaseType;
-                }
-            }
-            catch
-            {
-                // Best-effort: persistent data copy is non-critical
             }
         }
     }
